Application factory
Created a factory faction that can be explicitly involked from the script giving the script time to set configuration as well as create different instances of the application. The constructor imports most of the Flask extensions currently in use but because there is no application instance to initialize them with,it creates them uninitialized by passing no arguement to their constructors. The create_app function is the application factorywhic takes an arguement the name of a configuration to use for the application. The configurations are stored in config.py that can be imported directly into the application using the from_object() method available in Flask's app.config configuration object.